Move `temp_repo` fixture to vcs/git
